fixed up error with js-slice component
